// backups/release-candidates/zenova-definitiva-2026-03-01_1437/types.ts
export type WorkZone = 
  | 'URGENCES' 
  | 'TRAVAIL' 
  | 'URGENCES_TARDE'
  | 'TRAVAIL_TARDE'
  | 'ADMIN' 
  | 'ADM_PLUS'
  | 'TW' 
  | 'TW_ABROAD'
  | 'F' 
  | 'STRASBOURG' 
  | 'LIBERO'
  | 'RECUP'
  | 'FP'
  | 'CS'
  | 'SICK_LEAVE'
  | 'CA'
  | 'VACCIN'
  | 'VACCIN_AM'
  | 'VACCIN_PM'
  | 'VACCIN_PM_PLUS'
  | 'TRAVAIL_C'  // Complemento clínico corto
  | 'URGENCES_C'; // Complemento clínico corto

export interface Shift {
  id: WorkZone;
  label: string;
  description: string;
  color: string;
  textColor: string;
}

export interface Nurse {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  password?: string;
  order: number;
  mustChangePassword?: boolean;
  passwordResetRequired?: boolean;
}

export type UserRole = 'admin' | 'nurse' | 'viewer';
// ...
export interface ShiftCounts {
    TRAVAIL: number;
    TRAVAIL_TARDE: number;
    URGENCES: number;
    URGENCES_TARDE: number;
    ADMIN: number;
    ADM_PLUS: number;
    TW: number;
    TW_ABROAD: number;
    CA: number;
    FP: number;
    CS: number;
    RECUP: number;
    SICK_LEAVE: number;
    STRASBOURG: number;
    LIBERO: number;
    VACCIN: number;
    VACCIN_AM: number;
    VACCIN_PM: number;
    VACCIN_PM_PLUS: number;
}

export interface BalanceData {
    nurseId: string;
    monthlyCounts: ShiftCounts;
    annualCounts: ShiftCounts;
    monthlyTotalWorkDays: number;
    annualTotalWorkDays: number;
    monthlyTotalHours: number;
    annualTotalHours: number;
    monthlyTargetHours: number;
    annualTargetHours: number;
    monthlyBalance: number;
    annualBalance: number;
    hasConsecutiveAdmTw: boolean;
}